Conditions and bonus/reduction damage calculation variables have been moved from boardslots to a new Status class. This was done as only the active pokemon uses those effects, so it didn't make sense to store 6 instances when only one is required.

+++ GameHub/Core/GameCore/include/game.h
@@ -19,7 +19,7 @@ public:
   Game clone() const;
   void init(const CardFactory &_factory, GuiModule *const _drawer, const std::string &_deckA, const std::string &_deckB);
 
-  void dealDamage(const int _damage, const unsigned _id = 0);
+  void dealDamage(const int _damage, const size_t _id = 0);
   void addDamageCounter(const int _damage, const PTCG::PLAYER _player = PTCG::PLAYER::ENEMY, const unsigned _id = 0);
   void healDamage(const int _heal, const unsigned _id = 0);
   unsigned flipCoin(const unsigned _num);
